CT-9528 display agendas on calendar days
Introduces a new feature to display agenda indicators on the calendar days. This provides users with a visual cue for dates that have scheduled events or agendas.

A new `@Input()` property, `agendas`, has been added to the `owl-date-time` component. This property accepts an array of `CalendarAgenda` objects, where each object has a `start` and `end` date.

The calendar view now renders small dots below the date number for each agenda item on that day, indicating the presence of an event.

The color of the agenda indicators can be customized via the new CSS custom property `--owl-dt-agenda-color`.